+++ lib/ui/service_colors.dart
@@ -4,7 +4,10 @@ class ServiceColors {
   static const Color primary = Color(0xFFFFC700);
   static const Color primaryLight = Color(0xFFFFECBD);
   static const Color secondary = Color(0xFFD9D9D9);
-  static const Color cautionRed = Color(0x80FF0000);
   static const Color textColor = Color(0xFF333333);
   static const Color secondaryText = Color(0xFF797979);
+
+  static const Color cautionRed = Color(0x80FF0000);
+  static const Color wrongRed = Color(0xFFD70000);
+  static const Color correctGreen = Color(0xFF00D715);
 }
